pyproject.toml
[project]
name = "stt-tts-pipeline"
version = "0.1.0"
description = "Real-time STT→LLM→TTS pipeline with FastAPI and NeMo 2.0"
readme = "README.md"
requires-python = ">=3.10.19"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "websockets>=12.0",
    "python-multipart>=0.0.9",
    "nemo_toolkit[asr,tts]>=2.0.0",
    "torch>=2.1.0",
    "torchaudio>=2.1.0",
    "numpy>=1.24.0",
    "scipy>=1.11.0",
    "soundfile>=0.12.1",
    "librosa>=0.10.1",
    "silero-vad>=5.1",
    "vllm>=0.6.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "aiofiles>=23.2.1",
]
